// apps/git-client/src/generated/MultiRootResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MultiRootOutcome } from "./MultiRootOutcome";
import type { MultiRootRollbackStep } from "./MultiRootRollbackStep";

export type MultiRootResult = { outcomes: Array<MultiRootOutcome>, rollbackPlan: Array<MultiRootRollbackStep>, };