// apps/git-client/src/generated/MultiRootRollbackStep.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GitOperation } from "./GitOperation";
import type { RepositoryId } from "./RepositoryId";

export type MultiRootRollbackStep = { repositoryId: RepositoryId, path: string, description: string, operations: Array<GitOperation>, };